+++ src/main/java/io/xlorey/FluxLoader/interfaces/IControlsWidget.java
@@ -1,0 +1,11 @@
+package io.xlorey.FluxLoader.interfaces;
+
+/**
+ * Class template for implementing plugin controls
+ */
+public interface IControlsWidget {
+    /**
+     * Rendering controls in the plugin configuration menu
+     */
+    void render();
+}